% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_radar.R
\name{add_new_radar}
\alias{add_new_radar}
\title{Add new radar site.}
\usage{
add_new_radar(config_dir, radar)
}
\arguments{
\item{config_dir}{character, full path to BioConfigRadar.}

\item{radar}{named list, list(name, longitude, latitude, altitude, wavelength)}
}
\description{
Populate \code{radar_table} with new radar.
}
