% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_vp.R
\name{compute_vp}
\alias{compute_vp}
\title{Compute vertical profile.}
\usage{
compute_vp(radar_file, config_dir, rwanda = TRUE, radar_id = 1)
}
\arguments{
\item{radar_file}{character, full path to a single radar polar volume file.
The file data format should be \code{ODIM HDF5} or \code{NEXRAD Level 2}.}

\item{config_dir}{character, full path to BioConfigRadar.}

\item{rwanda}{logical, \code{TRUE} if the file is from MeteoRwanda new format \code{ODIM HDF5}.}

\item{radar_id}{integer, the radar id.}
}
\description{
Compute a vertical profile of bird and insect from a polar volume.
}
