% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodbradar.R
\name{create_bioDBConnection}
\alias{create_bioDBConnection}
\title{Create the authentication to BioDBRadar.}
\usage{
create_bioDBConnection(
  config_dir,
  db_name,
  username,
  password,
  host_name = "localhost",
  file_name = "biodb.con"
)
}
\arguments{
\item{config_dir}{character, full path to \code{BioConfigRadar}.}

\item{db_name}{character, PostgreSQL database name.}

\item{username}{character, PostgreSQL user name.}

\item{password}{character, PostgreSQL user password.}

\item{host_name}{character, the host name, default \code{"localhost"}.}

\item{file_name}{character, the file name of the connection, default \code{"biodb.con"}.}
}
\description{
Create the authentication to BioDBRadar.
}
