% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodbradar.R
\name{create_bioDBRadar}
\alias{create_bioDBRadar}
\title{Create BioDBRadar tables.}
\usage{
create_bioDBRadar(config_dir)
}
\arguments{
\item{config_dir}{character, full path to BioConfigRadar.}
}
\description{
Create the tables of BioDBRadar.
}
