% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vp.R
\name{get_vp}
\alias{get_vp}
\title{Vertical profile.}
\usage{
get_vp(data_info, query)
}
\arguments{
\item{data_info}{named list, list containing the vertical profiles dataset information.
To be replaced by a connection to postgresql.}

\item{query}{named list, user query \code{list(parameter, time)}.}
}
\value{
A named list, \code{list(status, data)}
}
\description{
Get a vertical profile for one datetime.
}
