% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vp.R
\name{get_vtip}
\alias{get_vtip}
\title{Vertical and time integration of profiles.}
\usage{
get_vtip(data_dir, query)
}
\arguments{
\item{data_dir}{character, full path to the directory containing the vertical profiles data.
To be replaced by a connection to postgresql.}

\item{query}{named list, user query \code{list(parameter, startTime, endTime)}.}
}
\value{
A named list, \code{list(status, data)}
}
\description{
Get the vertical and time integration of profiles.
}
