% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vp.R
\name{get_vtip_image}
\alias{get_vtip_image}
\title{Image vertical and time integration of profiles.}
\usage{
get_vtip_image(data_info, query)
}
\arguments{
\item{data_info}{named list, list containing the vertical profiles dataset information.
To be replaced by a connection to postgresql.}

\item{query}{named list, user query \code{list(parameter, startTime, endTime)}.}
}
\value{
A named list, \code{list(status, data)}
}
\description{
Plot the vertical and time integration of profiles.
}
