% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_vids.R
\name{process_rwanda_vids}
\alias{process_rwanda_vids}
\title{Spatial estimates of VID for MeteoRwanda radar data, multiple files.}
\usage{
process_rwanda_vids(
  bioradar_dir,
  start_time,
  end_time,
  radar_id = 1,
  reverse = FALSE
)
}
\arguments{
\item{bioradar_dir}{character, full path to the parent folder containing \code{BioConfigRadar}.}

\item{start_time}{character, start time to process, format 'yyyy-mm-dd hh:mm:ss'.}

\item{end_time}{character, end time to process, format 'yyyy-mm-dd hh:mm:ss'.}

\item{radar_id}{integer, the radar id.}

\item{reverse}{logical, if \code{TRUE} the data is computed from the last date in backward.}
}
\description{
Compute the spatial estimates of 
vertically integrated density of bird and insect
from MeteoRwanda new format \code{ODIM HDF5}
and write the outputs to a netCDF file.
}
