% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_vps.R
\name{process_rwanda_vps}
\alias{process_rwanda_vps}
\title{Vertical profile for MeteoRwanda radar data, multiple files.}
\usage{
process_rwanda_vps(bioradar_dir, start_time, end_time, radar_id = 1)
}
\arguments{
\item{bioradar_dir}{character, full path to the parent folder containing \code{BioConfigRadar}.}

\item{start_time}{character, start time to process, format 'yyyy-mm-dd hh:mm:ss'.}

\item{end_time}{character, end time to process, format 'yyyy-mm-dd hh:mm:ss'.}

\item{radar_id}{integer, the radar id.}
}
\description{
Compute a vertical profile of bird and insect
from MeteoRwanda new format \code{ODIM HDF5}
and store the outputs into \code{BioDBRadar}.
}
