% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_vids.R
\name{production_rwanda_vid}
\alias{production_rwanda_vid}
\title{Spatial estimates of VID for MeteoRwanda radar data, production mode.}
\usage{
production_rwanda_vid(bioradar_dir, radar_id = 1)
}
\arguments{
\item{bioradar_dir}{character, full path to the parent folder containing \code{BioConfigRadar}.}

\item{radar_id}{integer, the radar id.}
}
\description{
Compute the spatial estimates of 
vertically integrated density of bird and insect
from MeteoRwanda new format \code{ODIM HDF5}
and write the outputs to a netCDF file.
}
