% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_vps.R
\name{production_rwanda_vp}
\alias{production_rwanda_vp}
\title{Vertical profile for MeteoRwanda radar data, production mode.}
\usage{
production_rwanda_vp(bioradar_dir, radar_id = 1)
}
\arguments{
\item{bioradar_dir}{character, full path to the parent folder containing \code{BioConfigRadar}.}

\item{radar_id}{integer, the radar id.}
}
\description{
Compute a vertical profile of bird and insect
from MeteoRwanda new format \code{ODIM HDF5}
and store the outputs into \code{BioDBRadar}.
}
