% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodbradar.R
\name{verify_bioDBRadar}
\alias{verify_bioDBRadar}
\title{Verify BioDBRadar tables.}
\usage{
verify_bioDBRadar(config_dir)
}
\arguments{
\item{config_dir}{character, full path to BioConfigRadar.}
}
\description{
Check the tables and indexes of BioDBRadar.
}
