% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_vid.R
\name{wrapper_rwanda_vid}
\alias{wrapper_rwanda_vid}
\title{Spatial estimates of VID.}
\usage{
wrapper_rwanda_vid(radar_file, bioradar_dir, radar_id = 1)
}
\arguments{
\item{radar_file}{character, full path to a single radar polar volume file.
The file data format should be \code{ODIM HDF5} or \code{NEXRAD Level 2}.}

\item{bioradar_dir}{character, full path to the parent folder containing \code{BioConfigRadar}.}

\item{radar_id}{integer, the radar id.}
}
\description{
Compute the spatial estimates of 
vertically integrated density of bird and insect
from MeteoRwanda new format \code{ODIM HDF5}
and write the outputs to a netCDF file.
}
