% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwanda_vp.R
\name{wrapper_rwanda_vp}
\alias{wrapper_rwanda_vp}
\title{Process vertical profile for MeteoRwanda radar data.}
\usage{
wrapper_rwanda_vp(radar_file, bioradar_dir, radar_id = 1)
}
\arguments{
\item{radar_file}{character, full path to a single radar polar volume file.
The file data format should be \code{ODIM HDF5} or \code{NEXRAD Level 2}.}

\item{bioradar_dir}{character, full path to the parent folder containing \code{BioConfigRadar}.}

\item{radar_id}{integer, the radar id.}
}
\description{
Compute a vertical profile of bird and insect
from MeteoRwanda new format \code{ODIM HDF5}
and store the outputs into \code{BioDBRadar}.
}
