% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vp.R
\name{get_vp}
\alias{get_vp}
\title{Vertical profile.}
\usage{
get_vp(config_dir, query)
}
\arguments{
\item{config_dir}{character, full path to \code{BioConfigRadar}.}

\item{query}{named list, user query \code{list(parameter, time, species, radarID)}.}
}
\value{
A named list, \code{list(status, data)}
}
\description{
Get a vertical profile for one datetime.
}
