% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vp.R
\name{get_vpts}
\alias{get_vpts}
\title{Vertical profile time series.}
\usage{
get_vpts(config_dir, query)
}
\arguments{
\item{config_dir}{character, full path to \code{BioConfigRadar}.}

\item{query}{named list, user query \code{list(parameter, startTime, endTime, species, radarID)}.}
}
\value{
A named list, \code{list(status, data)}
}
\description{
Get a vertical profile time series.
}
