% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vp.R
\name{get_vtip_image}
\alias{get_vtip_image}
\title{Image vertical and time integration of profiles.}
\usage{
get_vtip_image(config_dir, query)
}
\arguments{
\item{config_dir}{character, full path to \code{BioConfigRadar}.}

\item{query}{named list, user query \code{list(parameter, startTime, endTime, species, radarID)}.}
}
\value{
A named list, \code{list(status, data)}
}
\description{
Plot the vertical and time integration of profiles.
}
